Listen for Pusher errors via a port
I don't know how to do an automated test of this. Change options.cluster
to an incorrect choice in the options parameter in the JavaScript call

    new Pusher(APP_KEY, options)

call to manually trigger an error. *But* -- the cluster has to be a valid
Pusher cluster.  If the cluster contains your app, you'll see no errors.
If it's a valid Pusher cluster that *doesn't* contain your app, the
error will get to Elm. If it's not a valid cluster, there will be
JavaScript errors about cross-site scripting that Elm never sees.