Ignore errors that don't have an error.data field.
When the internet connection is down, the error value passed to the
`pusher.connection.bind("error"...)` callback doesn't have an error.data
field. Maybe we should pass on something like "Unknown error", instead
of ignoring the problem, but in the case of a lost connection Pusher's
state_change captures the problem, changing from `connected` to
`connecting` to `unavailable`. (It does the reverse when connection is
restored.)