Fix (I hope!) Netlify deploy error
Netlify deploy said:

A Netlify Function is using "pusher" but that dependency has not been
installed yet.
By default, dependencies inside a Netlify Function's "package.json" are
not automatically installed.
There are several ways to fix this problem:
- Removing your Function's "package.json" and adding the dependencies to
the project's top-level "package.json" instead. This is the fastest and
safest solution.
- [other possible solutions]

I'm trying the first suggestion